Added more tests / refactoring.
